# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CropProfile:
    properties:
      id:
        type: integer
      ideal_season:
        example: summer
        type: string
      max_precipitation:
        example: 800
        type: number
      max_temp:
        example: 32
        type: number
      min_precipitation:
        example: 500
        type: number
      min_temp:
        example: 21
        type: number
      name:
        example: Milho
        type: string
    type: object
  models.CropRecommendation:
    properties:
      crop_name:
        example: Milho
        type: string
      justification:
        example: Temperatura e precipitação adequados
        type: string
      match_percentage:
        example: 95.5
        type: number
    type: object
  models.HarvestPlan:
    properties:
      precipitation:
        type: number
      recommendation:
        example: Colheita recomendada. Condições ideais de temperatura e sem chuvas.
        type: string
      temperature:
        type: number
      viability_score:
        description: 0 a 100
        example: 85
        type: integer
    type: object
  models.HealthResponse:
    properties:
      api:
        example: UP
        type: string
      database:
        example: UP
        type: string
      timestamp:
        example: "2026-06-06T11:20:00Z"
        type: string
      uptime:
        example: 14m3s
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: strongpassword
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.OpenMeteoResponse:
    properties:
      current:
        properties:
          precipitation:
            type: number
          temperature_2m:
            type: number
          time:
            type: string
          wind_speed_10m:
            type: number
        type: object
      latitude:
        type: number
      longitude:
        type: number
      timezone:
        type: string
    type: object
  models.ProblemDetail:
    properties:
      business_error:
        example: true
        type: boolean
      detail:
        example: A latitude fornecida não é um número válido.
        type: string
      instance:
        example: /api/v1/protected/engine/harvest
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Parâmetros Inválidos
        type: string
      type:
        example: https://api.agri-ai.com/errors/invalid-parameters
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: strongpassword
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.RiskAnalysis:
    properties:
      overall_risk_score:
        example: 45
        type: integer
      risk_factors:
        example:
        - Risco de Geada
        items:
          type: string
        type: array
    type: object
  models.WeatherCache:
    properties:
      created_at:
        type: string
      data:
        description: JSON string da resposta
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      query_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API backend for agricultural AI insights
  title: Agri-AI API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica e-mail/senha e retorna um JWT
      parameters:
      - description: Credenciais do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token JWT
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Requisição inválida
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Credenciais inválidas
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login do usuário
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma nova conta com e-mail e senha
      parameters:
      - description: Credenciais do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuário criado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Requisição inválida
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Usuário já existe
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar usuário
      tags:
      - auth
  /healthz:
    get:
      description: Verifica status da API e conexões com o banco de dados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.HealthResponse'
      summary: Liveness/Readiness Probe
      tags:
      - system
  /protected/crops:
    get:
      consumes:
      - application/json
      description: Retorna todos os perfis de culturas (Crops) disponíveis no banco
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CropProfile'
            type: array
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/models.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Listar Culturas Agrícolas
      tags:
      - crops
  /protected/engine/crop-selector:
    get:
      consumes:
      - application/json
      description: Faz um match climático entre as safras do banco e o clima da região
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      - description: 'Estação do ano (ex: summer, winter)'
        in: query
        name: season
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CropRecommendation'
            type: array
        "400":
          description: Parâmetros inválidos
          schema:
            $ref: '#/definitions/models.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/models.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Seleção de Cultivo
      tags:
      - engine
  /protected/engine/harvest:
    get:
      consumes:
      - application/json
      description: Avalia as condições climáticas e indica a viabilidade de colheita
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HarvestPlan'
        "400":
          description: Parâmetros inválidos
          schema:
            $ref: '#/definitions/models.ProblemDetail'
        "429":
          description: Muitas requisições (Rate Limit)
          schema:
            $ref: '#/definitions/models.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/models.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Predição de Colheita
      tags:
      - engine
  /protected/engine/risk-analysis:
    get:
      consumes:
      - application/json
      description: Avalia risco baseado em temperatura, vento e chuva
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RiskAnalysis'
        "400":
          description: Parâmetros inválidos
          schema:
            $ref: '#/definitions/models.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/models.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Análise de Risco (Geada / Incêndio)
      tags:
      - engine
  /protected/weather:
    get:
      consumes:
      - application/json
      description: Retorna a temperatura e precipitação de uma coordenada. Utiliza
        cache no banco de dados e Open-Meteo como fallback.
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lon
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OpenMeteoResponse'
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno ao buscar clima
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Buscar Clima Atual
      tags:
      - weather
  /protected/weather/cache:
    get:
      consumes:
      - application/json
      description: Retorna todos os dados cacheados de clima com suporte a paginação
      parameters:
      - description: Limite de resultados (default 10)
        in: query
        name: limit
        type: integer
      - description: Deslocamento de resultados (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.WeatherCache'
            type: array
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/models.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Listar Cache Climático
      tags:
      - weather
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"